LayoutTests/accessibility/mac/accesskey-expected.txt

webkit

This test makes sure that elements return their accesskeys.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS link.allAttributes().indexOf(linkAccessKey) != -1 is true
PASS text.allAttributes().indexOf(inputAccessKey) != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
